rename worker block files batches
